+++ source/extensions/isaacsim.code_editor.vscode/isaacsim/code_editor/vscode/tests/__init__.py
@@ -13,7 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Tests for the Visual Studio Code integration functionality."""
-
+"""Tests for the VS Code integration extension."""
 
 scan_for_test_modules = True